Fix gitversion.h gen for git worktrees. (#10)
* Use git rev-parse HEAD in gitversion.

Allow using of git worktrees for cflib packages. The .git is a file in a
worktree, therefore git rev-parse HEAD is used to determine the hash.

* Remove depend from gitversion.

Using git worktrees, .git is a file and not a directory. Therefore
a dependency on it for detecting hash changes does not work and the
code is no longer functional.

---------

Co-authored-by: Andre Vehreschild <vehre@badgersystems.de>